fix: make video runtime installs POST-only (#4895)
