+++ server/routes/videoGen.test.js
@@ -36,8 +36,17 @@ vi.mock('../lib/childProcess.js', async (importOriginal) => ({
 // child as a clean exit, i.e. "capable". Pin the verdict explicitly so each
 // test states the capability it is exercising.
 const loraCapability = vi.hoisted(() => ({ capable: false }));
+const runtimeProbes = vi.hoisted(() => ({
+  isByovRuntimeInstalled: vi.fn(() => false),
+  isByovRuntimeReady: vi.fn(async () => false),
+  isByovRuntimeCurrent: vi.fn(async () => false),
+  invalidateByovReadyCache: vi.fn(),
+  invalidateByovLoraCapabilityCache: vi.fn(),
+  invalidateRuntimeFingerprintCache: vi.fn(),
+}));
 vi.mock('../services/videoGen/runtimes.js', async (importOriginal) => ({
   ...(await importOriginal()),
+  ...runtimeProbes,
   resolveByovRuntimeLoraCapable: vi.fn(async (runtime) => runtime === 'minimax_h3' && loraCapability.capable),
 }));
 
@@ -116,12 +125,7 @@ vi.mock('../services/videoGen/local.js', () => ({
     },
     hunyuan: { id: 'hunyuan', label: 'HunyuanVideo MLX', venvPython: '/tmp/hunyuan.py', installEnvVar: 'INSTALL_HUNYUAN', repoUrl: 'x', repoDir: '/tmp' },
   },
-  isByovRuntimeInstalled: vi.fn(() => false),
-  isByovRuntimeReady: vi.fn(async () => false),
-  isByovRuntimeCurrent: vi.fn(async () => false),
-  invalidateByovReadyCache: vi.fn(),
-  invalidateByovLoraCapabilityCache: vi.fn(),
-  invalidateRuntimeFingerprintCache: vi.fn(),
+  ...runtimeProbes,
   // /status now surfaces a runtime block (host chip/os + per-runtime versions).
   // Mock returns a fixed shape so the status test can assert it's wired through.
   resolveRuntimeFingerprint: vi.fn(async () => ({
@@ -301,6 +305,9 @@ describe('videoGen routes', () => {
     // bailed before the route consumed it can't leak into the next test.
     pendingUpload.current = null;
     installProcess.spawn.mockReset().mockImplementation(() => installProcess.makeChild());
+    runtimeProbes.isByovRuntimeInstalled.mockReturnValue(false);
+    runtimeProbes.isByovRuntimeReady.mockResolvedValue(false);
+    runtimeProbes.isByovRuntimeCurrent.mockResolvedValue(false);
   });
 
   describe('GET /status', () => {
@@ -429,12 +436,22 @@ describe('videoGen routes', () => {
     });
   });
 
-  describe('GET /setup/runtime-install', () => {
+  describe('/setup/runtime-install', () => {
+    it('keeps GET read-only and returns the current runtime status', async () => {
+      const r = await request(app).get('/api/video-gen/setup/runtime-install?runtime=wan22');
+      expect(r.status).toBe(200);
+      expect(r.body).toMatchObject({
+        runtime: 'wan22', installed: false, binaryPresent: false,
+        packagesReady: false, current: false,
+      });
+      expect(installProcess.spawn).not.toHaveBeenCalled();
+    });
+
     it('short-circuits only when the runtime packages and pinned revision are current', async () => {
       videoGenService.isByovRuntimeInstalled.mockReturnValueOnce(true);
       videoGenService.isByovRuntimeReady.mockResolvedValueOnce(true);
       videoGenService.isByovRuntimeCurrent.mockResolvedValueOnce(true);
-      const r = await request(app).get('/api/video-gen/setup/runtime-install?runtime=wan22');
+      const r = await request(app).post('/api/video-gen/setup/runtime-install?runtime=wan22');
       expect(r.status).toBe(200);
       expect(r.text).toContain('"type":"complete"');
       expect(r.text).toContain('Already installed');
@@ -451,7 +468,7 @@ describe('videoGen routes', () => {
       videoGenService.isByovRuntimeCurrent
         .mockResolvedValueOnce(false)
         .mockResolvedValueOnce(true);
-      const r = await request(app).get('/api/video-gen/setup/runtime-install?runtime=wan22');
+      const r = await request(app).post('/api/video-gen/setup/runtime-install?runtime=wan22');
       expect(r.status).toBe(200);
       expect(r.text).toContain('"type":"complete"');
       expect(r.text).toContain('Wan 2.2 MLX ready');
